backend/requirements.txt
django~=3.2.0
django-filter~=22.1.0
djangorestframework~=3.12.0
djoser~=2.1.0
drf-yasg~=1.21.5
flake8~=5.0.0
flake8-broken-line~=0.6.0
flake8-isort~=5.0.0
flake8-return~=1.2.0
pep8-naming~=0.13.0
